Comit intermediario

+++ BPMNLetter/Yaoqiang/src/org/yaoqiang/bpmn/editor/action/ModelActions.java
@@ -91,9 +91,8 @@ import poli.mestrado.parser.bpmn2use.graph.Graph;
 import poli.mestrado.parser.bpmn2use.graph.GraphHelper;
 import poli.mestrado.parser.bpmn2use.graph.Vertice;
 import poli.mestrado.parser.bpmn2use.tag.ProcessDiagram;
-import poli.mestrado.parser.bpmn2use.tag.ProcessTag;
-import poli.mestrado.parser.bpmn2use.tag.task.AbstractTaskElement;
 import poli.mestrado.parser.uml2use.UmlFileManager;
+import poli.mestrado.parser.util.MyConstants;
 
 import com.mxgraph.canvas.mxICanvas;
 import com.mxgraph.io.mxCodec;
@@ -105,7 +104,6 @@ import com.mxgraph.util.mxDomUtils;
 import com.mxgraph.util.mxResources;
 import com.mxgraph.util.mxUtils;
 import com.mxgraph.util.mxXmlUtils;
-import com.sun.beans.decoder.DocumentHandler;
 
 /**
  * ModelActions
@@ -379,10 +377,69 @@ public class ModelActions extends AbstractAction {
 				editor.getUndoManager().clear();
 			}
 		} else if(EXPORT_USE.equals(type)){
-			try {
-				String path = 
-//						"C:/Users/Rafael/Desktop/Nova pasta (3)/";
-						UmlFileManager.getInstance().getExportXmiFile().getAbsolutePath().substring(0, UmlFileManager.getInstance().getExportXmiFile().getAbsolutePath().lastIndexOf(File.separator)+1);
+				String wd;
+				boolean dialogShown = false;
+				String filename = null;
+				String ext = null;
+				File bpmnImageFile = null;
+				FileFilter wordFileFilter = new DefaultFileFilter(".doc", "Word " + mxResources.get("file") + " (.doc)");
+				FileFilter htmlFileFilter = new DefaultFileFilter(".html", "HTML " + mxResources.get("file") + " (.html)");
+				FileFilter PdfFileFilter = new DefaultFileFilter(".pdf", "PDF " + mxResources.get("file") + " (.pdf)");
+				
+				if (EditorConstants.LAST_OPEN_DIR != null) {
+					wd = EditorConstants.LAST_OPEN_DIR;
+				} else if (BPMNEditor.getCurrentFile() != null) {
+					wd = BPMNEditor.getCurrentFile().getParent();
+				} else {
+					wd = System.getProperty("user.dir");
+				}
+
+				JFileChooser fc = new JFileChooser(wd);
+				fc.setAcceptAllFileFilterUsed(false);
+				if (BPMNEditor.getCurrentFile() != null) {
+					String name = BPMNEditor.getCurrentFile().getName();
+					if (name.endsWith(".bpmn20.xml")) {
+						name = name.substring(0, name.lastIndexOf(".bpmn20.xml"));
+					} else {
+						name = name.substring(0, name.lastIndexOf("."));
+					}
+					fc.setSelectedFile(new File(name));
+				}
+				fc.addChoosableFileFilter(wordFileFilter);
+				fc.addChoosableFileFilter(htmlFileFilter);
+				fc.addChoosableFileFilter(PdfFileFilter);
+				
+				int rc = fc.showDialog(null, "3".equals(stringValue) ? mxResources.get("savePNGfile") : mxResources.get("save"));
+				dialogShown = true;
+				if (rc != JFileChooser.APPROVE_OPTION) {
+					return;
+				} else {
+					EditorConstants.LAST_OPEN_DIR = fc.getSelectedFile().getParent();
+					Utils.saveToConfigureFile("lastOpenDir", EditorConstants.LAST_OPEN_DIR);
+				}
+				
+				filename = fc.getSelectedFile().getAbsolutePath();
+				ext = fc.getFileFilter().getDescription().substring(fc.getFileFilter().getDescription().lastIndexOf("(")+1, fc.getFileFilter().getDescription().lastIndexOf("")-1);
+				
+				String documentPath = "";
+				if(!filename.contains(".")){
+					documentPath = filename+ext; 
+				}else{
+					documentPath = filename;
+				}
+				
+				File documentFile = new File(documentPath);
+				if (documentFile.exists()
+						&& JOptionPane.showConfirmDialog(graphComponent, mxResources.get("overwriteExistingFile")) != JOptionPane.YES_OPTION) {
+					return;
+				}
+				
+				documentFile.delete();
+				
+				
+				try {
+				
+				String path = UmlFileManager.getInstance().getExportXmiFile().getAbsolutePath().substring(0, UmlFileManager.getInstance().getExportXmiFile().getAbsolutePath().lastIndexOf(File.separator)+1);
 				File bpmnFile = new File(path+"bpmnmodel.bpmn");
 				if(bpmnFile.exists()){
 					bpmnFile.delete();
@@ -392,51 +449,36 @@ public class ModelActions extends AbstractAction {
 				bpmnModel.setExporter("Yaoqiang BPMN Editor");
 				bpmnModel.setExporterVersion(BaseEditor.VERSION);
 				mxUtils.writeFile(mxXmlUtils.getXml(codec.encode().getDocumentElement()), bpmnFile.getAbsolutePath());
-
 				editor.setModified(false);
 				editor.setCurrentFile(bpmnFile);
 				BPMNEditorUtils.addRecentFiletoList(editor, bpmnFile.getAbsolutePath());
-				System.out.println("O arquivo BPMN foi salvo com sucesso!");
-//------------------------------------Comente daqui
-				Thread.sleep(3000);
 
-				ProcessDiagram processDiagram = ParserBPMNHelper.getInstance().getProcessDiagram();
-				System.out.println("\n--------------------------------resultado do parser------------------------------");
-				System.out.println(processDiagram.toString());
-				
-//				DocumentHelper.getInstance().geneareDocument(processDiagram);
-
-				Graph graph1 = GraphHelper.getInstance().generateGraph(processDiagram);
-				System.out.println("\n______________________________Gerando o grafo__________________________________________________________");
-				System.out.println(graph1.toString());
-
-
-				List<List<Vertice>>  pahts = GraphHelper.getInstance().generatePaths();
-				System.out.println("\n--------------------------------Camhinhos gerados--------------------------------------------------------");
-				int i = 1;
-				for (List<Vertice> list : pahts) {
-					System.out.println("\n-----------------------------Path"+i+"-------------------------------\n");
-					for (Vertice vertice : list) {
-//						if(vertice.getVeriticeElement() instanceof AbstractTaskElement){
-//							System.out.print(vertice.getVeriticeElement().getName()+"("+vertice.getVeriticeElement().getId()+") > ");
-//							System.out.print(vertice.getVeriticeElement().getName().replace("\n", " ")+" > ");
-//						}
-//						else{
-//							System.out.print(vertice.getVeriticeElement().getName()+"("+vertice.getVeriticeElement().getId()+") > ");
-//						}
-						System.out.print(vertice.getVeriticeElement().getName()+"("+vertice.getVeriticeElement().getId()+")"+"->");
+				BufferedImage image = mxCellRenderer.createBufferedImage(graph, null, 1, null, graphComponent.isAntiAlias(), null, false,
+						graphComponent.getCanvas());
+				if (image != null) {
+					String imagefilename = "";
+					if(!filename.contains(".")){
+						imagefilename = filename + ".png";
+					}else{
+						imagefilename = filename.substring(0,filename.lastIndexOf(".")) + ".png";
 					}
-					i++;
+					bpmnImageFile = new File(imagefilename);
+					if(bpmnImageFile.exists()){
+						bpmnImageFile.delete();
+					}
+					ImageIO.write(image, "png", new File(imagefilename));
+					editor.getFragmentsPalette().addFragmentTemplate(new File(imagefilename));
+					editor.getFragmentsPalette().setPreferredCols(1, 195 * image.getHeight() / image.getWidth());
 				}
 				
-//				String path1 = UmlFileManager.getInstance().getExportXmiFile().getAbsolutePath().substring(0, UmlFileManager.getInstance().getExportXmiFile().getAbsolutePath().lastIndexOf(File.separator)+1)+"bpmnPathsModel.serializabe";
-//			        GenereteSerializable.saveModel(path1, pahts);
-//				
-//				if(poli.mestrado.parser.util.Constants.onlyMainInAsslScript){
-//					GenerateAsslScriptsManager.getInstance().generateAsslScriptOnlyMain(pahts);
-//				}else{
-//					GenerateAsslScriptsManager.getInstance().generateAsslScriptsManiAndMethod(pahts);
-//				}
+				System.out.println("O arquivo BPMN foi salvo com sucesso!");
+				//------------------------------------Comente daqui
+				Thread.sleep(3000);
+				ProcessDiagram processDiagram = ParserBPMNHelper.getInstance().getProcessDiagram();
+				
+				DocumentHelper.getInstance().geneareDocument(documentPath, processDiagram);
+
+				JOptionPane.showMessageDialog(graphComponent, "A carta foi salva em: "+documentPath, mxResources.get("warning"), JOptionPane.WARNING_MESSAGE);
 				//------------------------------------ATE aqui Comente daqui
 			} catch (Exception ex) {
 				ex.printStackTrace();
@@ -997,7 +1039,7 @@ public class ModelActions extends AbstractAction {
 			} else if (el instanceof Participant) {
 				el = bpmnModel.getProcess(((Participant) el).getProcessRef());
 			}
-			if (el instanceof Activity || el instanceof BPMNProcess) {
+			if (el instanceof Activity || el instanceof Gateway || el instanceof Event || el instanceof BPMNProcess) {
 				dialog.initDialog().editBPMNElement(((XMLComplexElement) el).get("resources"));
 			}
 		} else if (DEPLOY.equals(type)) {
